Cleaned up MainActivity, migrated code to AuthenticationHandler
