+++ app/src/main/java/com/justcorrections/grit/MainActivity.java
@@ -3,23 +3,13 @@ package com.justcorrections.grit;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.Snackbar;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.TextView;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.justcorrections.grit.modules.account.AccountFragment;
+import com.justcorrections.grit.auth.AuthenticationHandler;
 import com.justcorrections.grit.modules.account.LoginFragment;
-import com.justcorrections.grit.modules.account.OnAccountRequestListener;
 import com.justcorrections.grit.modules.map.MapFragment;
 import com.justcorrections.grit.modules.map.ResourceDetailFragment;
 import com.justcorrections.grit.modules.mystery.MysteryFragment;
@@ -29,15 +19,14 @@ import com.justcorrections.grit.utils.NavigationHandler;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MainActivity extends AppCompatActivity implements OnAccountRequestListener {
+public class MainActivity extends AppCompatActivity {
 
     private NavigationHandler navigationHandler;
+    private AuthenticationHandler authHandler;
     private FirebaseAuth auth;
     private FirebaseAuth.AuthStateListener authStateListener;
     private BottomNavigationView bottomNav;
-    private CoordinatorLayout snackbarView;
     private List<Fragment> fragments = new ArrayList<>();
-    private AlertDialog status;
 
     private DatabaseHelper helper;
 
@@ -46,43 +35,10 @@ public class MainActivity extends AppCompatActivity implements OnAccountRequestL
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        bottomNav = (BottomNavigationView) findViewById(R.id.main_bottom_nav);
-        snackbarView = (CoordinatorLayout) findViewById(R.id.main_snackbar_view);
-
+        bottomNav = findViewById(R.id.main_bottom_nav);
         navigationHandler = new NavigationHandler(this);
-
         helper = DatabaseHelper.getInstance();
-
-        auth = FirebaseAuth.getInstance();
-        // onAuthStateChanged gets called when AuthStateListener is registered
-        authStateListener = new FirebaseAuth.AuthStateListener() {
-            @Override
-            public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
-                Log.d("Main Activity", "onAuthStateChanged: user state changed");
-                // user logged in
-                if (firebaseAuth.getCurrentUser() != null) {
-                    fragments.remove(1);
-                    fragments.add(1, AccountFragment.newInstance());
-                }
-                // user logged out
-                else {
-                    fragments.remove(1);
-                    fragments.add(1, LoginFragment.newInstance(null));
-                }
-
-                // TODO: add a transition using anim resource files
-                if (bottomNav.getSelectedItemId() == R.id.menu_account) {
-                    getSupportFragmentManager()
-                            .beginTransaction()
-                            .replace(R.id.main_fragment_holder, fragments.get(1))
-                            .commit();
-                }
-
-            }
-        };
-
-        auth.addAuthStateListener(authStateListener);
-
+        authHandler = new AuthenticationHandler(this, this.fragments);
         setupFragments();
 
         bottomNav.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
@@ -111,7 +67,7 @@ public class MainActivity extends AppCompatActivity implements OnAccountRequestL
         navigationHandler.onBackPressed();
     }
 
-    private void setSelected(int position) {
+    public void setSelected(int position) {
         if (position == 0) {
             navigateTo(MapFragment.newInstance());
         } else if (position == 1) {
@@ -127,83 +83,8 @@ public class MainActivity extends AppCompatActivity implements OnAccountRequestL
         fragments.add(2, MysteryFragment.newInstance("Parameter 1", "Parameter 2"));
     }
 
-    @Override
-    public void onResetRequest(String email) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        status = builder.setView(R.layout.dialog_status)
-                .setCancelable(false)
-                .create();
-        status.show();
-
-        auth.sendPasswordResetEmail(email).addOnCompleteListener(new OnCompleteListener<Void>() {
-            @Override
-            public void onComplete(@NonNull Task<Void> task) {
-                if (task.isSuccessful()) {
-                    status.findViewById(R.id.dialog_success).setVisibility(View.VISIBLE);
-                    ((TextView) status.findViewById(R.id.dialog_status)).setText(R.string.reset_email_sent);
-                } else {
-                    status.findViewById(R.id.dialog_error).setVisibility(View.VISIBLE);
-                    ((TextView) status.findViewById(R.id.dialog_status)).setText(R.string.reset_email_failed);
-                }
-                status.findViewById(R.id.dialog_progress).setVisibility(View.INVISIBLE);
-                status.setCancelable(true);
-            }
-        });
-    }
-
-    @Override
-    public void onLoginRequest(String email, String password) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        status = builder.setView(R.layout.dialog_status)
-                .setCancelable(false)
-                .create();
-        status.show();
-        ((TextView) status.findViewById(R.id.dialog_status)).setText(R.string.signing_in);
-        auth.signInWithEmailAndPassword(email, password).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-            @Override
-            public void onComplete(@NonNull Task<AuthResult> task) {
-                if (task.isSuccessful()) {
-                    status.dismiss();
-                } else {
-                    status.findViewById(R.id.dialog_error).setVisibility(View.VISIBLE);
-                    ((TextView) status.findViewById(R.id.dialog_status)).setText(R.string.invalid_credentials);
-                    status.findViewById(R.id.dialog_progress).setVisibility(View.INVISIBLE);
-                    status.setCancelable(true);
-                }
-            }
-        });
-    }
-
-    @Override
-    public void onCreateRequest(String email, String password) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        status = builder.setView(R.layout.dialog_status)
-                .setCancelable(false)
-                .create();
-        status.show();
-        ((TextView) status.findViewById(R.id.dialog_status)).setText(R.string.creating_account);
-        auth.createUserWithEmailAndPassword(email, password).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-            @Override
-            public void onComplete(@NonNull Task<AuthResult> task) {
-                if (task.isSuccessful()) {
-                    status.dismiss();
-                    fragments.remove(1);
-                    fragments.add(1, AccountFragment.newInstance());
-                    setSelected(1);
-                } else {
-                    status.findViewById(R.id.dialog_error).setVisibility(View.VISIBLE);
-                    ((TextView) status.findViewById(R.id.dialog_status)).setText(R.string.error_occurred);
-                    status.findViewById(R.id.dialog_progress).setVisibility(View.INVISIBLE);
-                    status.setCancelable(true);
-                }
-            }
-        });
-    }
-
-    @Override
-    public void onFail(String message) {
-        Snackbar snackbar = Snackbar.make(snackbarView, message, Snackbar.LENGTH_LONG);
-        snackbar.show();
+    public AuthenticationHandler getAuthHandler() {
+        return this.authHandler;
     }
 
     public void showResourceDetailFragment(String resourceID) {
